pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ehnn-crispr"
version = "0.1.1"
description = "EHNN – End-to-End CRISPR Off-Target Predictor: feature extraction → normalization → hybrid encoding → model training"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [{ name = "akhtar Sayed" }]
keywords = ["CRISPR", "off-target", "deep learning", "bioinformatics", "EHNN"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics"
]

dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "biopython>=1.83",
    "matplotlib>=3.7",
    "seaborn>=0.13",
    "typer>=0.9.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
ml = [
    "tensorflow>=2.13,<2.18",
]
dev = [
    "pytest>=7.4",
    "black>=23.7",
    "isort>=5.12",
    "flake8>=6.1",
    "mypy>=1.5",
    "pre-commit>=3.4",
]

[project.urls]
homepage = "https://github.com/akhtarsayed/EHNN-CRISPR-OffTarget-Prediction/"
repository = "https://github.com/akhtarsayed/EHNN-CRISPR-OffTarget-Prediction/"
issues = "https://github.com/akhtarsayed/EHNN-CRISPR-OffTarget-Prediction//issues"

[project.scripts]
ehnn = "ehnn.cli:main"

[tool.setuptools]
packages = ["ehnn"]
include-package-data = false

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
